--- models/http-error.js
@@ -1,8 +1,0 @@
-class HttpError extends Error {
-    constructor(message, errorCode) {
-      super(message);
-      this.code = errorCode;
-    }
-  }
-  
-  module.exports = HttpError;
