+++ models/http-errors.js
@@ -1,0 +1,8 @@
+class HttpError extends Error {
+  constructor(message, errorCode) {
+    super(message);
+    this.code = errorCode;
+  }
+}
+
+module.exports = HttpError;
