+++ routes/users-route.js
@@ -2,7 +2,7 @@ const express = require('express');
 const { check } = require('express-validator');
 
 const usersController = require('../controllers/users-controller');
-const fileUpload = require('../middleware/file-upload');
+// const fileUpload = require('../middleware/file-upload');
 
 const router = express.Router();
 
@@ -10,7 +10,6 @@ router.get('/', usersController.getUsers);
 
 router.post(
   '/signup',
-  fileUpload.single('image'),
   [
     [check('email').isEmail().normalizeEmail().trim()],
     check('username').not().isEmpty().trim(),
