+++ routes/videos-route.js
@@ -4,8 +4,6 @@ const { check } = require('express-validator');
 
 const videoController = require('../controllers/videos-controller');
 
-const fileUpload = require('../middleware/file-upload');
-
 const router = express.Router();
 
 router.get('/', videoController.getVideos);
@@ -16,7 +14,6 @@ router.get('/user/:uid', videoController.getVideosByUserId);
 
 router.post(
   '/',
-  fileUpload.single('image'),
   [
     check('title').not().isEmpty().trim(),
     check('description').isLength({ min: 5 }).not().isEmpty().trim(),
